Reorganized TextMessageReceivedEvents #252
